requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
dj-rest-auth==6.0.0
Django==4.2.13
django-admin-interface==0.28.8
django-allauth==0.63.3
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-filter==24.2
django-jazzmin==3.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt

exceptiongroup==1.2.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
Markdown==3.6
openai==1.35.3
packaging==24.1
pillow==10.4.0
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT

python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.0
text-unidecode==1.3
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
zipp==3.19.2
django-countries==7.6.1
django-admin-rangefilter==0.13.1
pandas==2.3.2
openpyxl==3.1.5
django-crispy-forms==2.4
crispy-bootstrap4==2025.6
firebase-admin
django-cors-headers==4.4.0
mysqlclient==2.2.4
coverage
Faker==19.13.0